Added navigator and moved jdtls config and start to navigator
